Add advanced session duration filter option (#1224)
* Add advanced session duration filter option

* Add empty placeholder

* Update type

+++ backend/model/model.go
@@ -471,8 +471,8 @@ type DateRange struct {
 }
 
 type LengthRange struct {
-	Min int
-	Max int
+	Min float64
+	Max float64
 }
 
 type UserProperty struct {
